Added support of rotations using turf and JIMP.
