+++ converters/kml/index.ts
@@ -277,26 +277,26 @@ export class KMLToGeoPackage {
         let geometryData;
         if (node.hasOwnProperty(KMLTAGS.GEOMETRY_TAGS.POLYGON)) {
           isGeom = true;
-          geometryData = this.handlePolygons(node);
+          geometryData = KMLUtilities.kmlPolygonToGeoJson(node);
         } else if (node.hasOwnProperty(KMLTAGS.GEOMETRY_TAGS.POINT)) {
           isGeom = true;
-          geometryData = this.handlePoints(node);
+          geometryData = KMLUtilities.kmlPointToGeoJson(node);
         } else if (node.hasOwnProperty(KMLTAGS.GEOMETRY_TAGS.LINESTRING)) {
           isGeom = true;
-          geometryData = this.handleLineStrings(node);
+          geometryData = KMLUtilities.kmlLineStringToGeoJson(node);
         } else if (node.hasOwnProperty(KMLTAGS.GEOMETRY_TAGS.MULTIGEOMETRY)) {
           isGeom = true;
           geometryData = { type: 'GeometryCollection', geometries: [] };
           if (node.MultiGeometry.hasOwnProperty(KMLTAGS.GEOMETRY_TAGS.POINT)) {
-            const temp = this.handlePoints(node.MultiGeometry);
+            const temp = KMLUtilities.kmlPointToGeoJson(node.MultiGeometry);
             geometryData['geometries'].push(temp);
           }
           if (node.MultiGeometry.hasOwnProperty(KMLTAGS.GEOMETRY_TAGS.LINESTRING)) {
-            const temp = this.handleLineStrings(node.MultiGeometry);
+            const temp = KMLUtilities.kmlLineStringToGeoJson(node.MultiGeometry);
             geometryData['geometries'].push(temp);
           }
           if (node.MultiGeometry.hasOwnProperty(KMLTAGS.GEOMETRY_TAGS.POLYGON)) {
-            const temp = this.handlePolygons(node.MultiGeometry);
+            const temp = KMLUtilities.kmlPolygonToGeoJson(node.MultiGeometry);
             geometryData['geometries'].push(temp);
           }
         }
@@ -683,96 +683,5 @@ export class KMLToGeoPackage {
     const rgb = abgr.slice(6, 8) + abgr.slice(4, 6) + abgr.slice(2, 4);
     const a = parseInt('0x' + abgr.slice(0, 2)) / 255;
     return { rgb, a };
-  }
-
-  /**
-   * Takes in a KML Point and returns a GeoJSON formatted object.
-   * @param node The data from xmlStream with the selector of Placemark.
-   */
-  private handlePoints(node: { Point }): { type: string; coordinates: number[] } {
-    let geometryData;
-    if (node[KMLTAGS.GEOMETRY_TAGS.POINT].length === 1) {
-      geometryData = { type: 'Point', coordinates: [] };
-    } else {
-      geometryData = { type: 'MultiPoint', coordinates: [] };
-    }
-    node[KMLTAGS.GEOMETRY_TAGS.POINT].forEach(point => {
-      const coordPoint = point.coordinates.split(',');
-      const coord = [parseFloat(coordPoint[0]), parseFloat(coordPoint[1])];
-      if (node.Point.length === 1) {
-        geometryData['coordinates'] = [parseFloat(coordPoint[0]), parseFloat(coordPoint[1])];
-      } else {
-        geometryData['coordinates'].push(coord);
-      }
-    });
-    return geometryData;
-  }
-
-  /**
-   * Takes in a KML LineString and returns a GeoJSON formatted object.
-   * @param node The data from xmlStream with the selector of Placemark.
-   */
-  private handleLineStrings(node: { LineString }): { type: string; coordinates: number[] } {
-    let geometryData;
-    if (node[KMLTAGS.GEOMETRY_TAGS.LINESTRING].length === 1) {
-      geometryData = { type: 'LineString', coordinates: [] };
-    } else {
-      geometryData = { type: 'MultiLineString', coordinates: [] };
-    }
-    node[KMLTAGS.GEOMETRY_TAGS.LINESTRING].forEach(element => {
-      const coordPoints = element.coordinates.split(' ');
-      const coordArray = [];
-      coordPoints.forEach(element => {
-        element = element.split(',');
-        coordArray.push([Number(element[0]), Number(element[1])]);
-      });
-      if (node[KMLTAGS.GEOMETRY_TAGS.LINESTRING].length === 1) {
-        geometryData['coordinates'] = coordArray;
-      } else {
-        geometryData['coordinates'].push(coordArray);
-      }
-    });
-    return geometryData;
-  }
-
-  /**
-   * Takes in a KML Polygon and returns a GeoJSON formatted object.
-   * @param node The data from xmlStream with the selector of Placemark.
-   */
-  private handlePolygons(node: { Polygon }): { type: string; coordinates: number[] } {
-    let geometryData;
-    if ([KMLTAGS.GEOMETRY_TAGS.POLYGON].length === 1) {
-      geometryData = { type: 'Polygon', coordinates: [] };
-    } else {
-      geometryData = { type: 'MultiPolygon', coordinates: [] };
-    }
-    node[KMLTAGS.GEOMETRY_TAGS.POLYGON].forEach(element => {
-      const coordText = element.outerBoundaryIs.LinearRing[0].coordinates;
-      const coordRing = coordText.split(' ');
-      const coordArray = [];
-      coordRing.forEach(element => {
-        element = element.split(',');
-        coordArray.push([parseFloat(element[0]), parseFloat(element[1])]);
-      });
-
-      const temp = [coordArray];
-      if (node[KMLTAGS.GEOMETRY_TAGS.POLYGON].hasOwnProperty('innerBoundaryIs')) {
-        const coordText = element.innerBoundaryIs.LinearRing[0].coordinates;
-        const coordRing = coordText.split(' ');
-        const coordArray = [];
-        coordRing.forEach(elementRing => {
-          elementRing = elementRing.split(',');
-          coordArray.push([parseFloat(elementRing[0]), parseFloat(elementRing[1])]);
-        });
-        temp.push(coordArray);
-      }
-
-      if (node[KMLTAGS.GEOMETRY_TAGS.POLYGON].length === 1) {
-        geometryData['coordinates'] = temp;
-      } else {
-        geometryData['coordinates'].push(temp);
-      }
-    });
-    return geometryData;
   }
 }
